Rename package so src/test match
